requirements.txt
fastapi
httpx==0.28.1
websockets==12.0
protobuf==5.29.3
rich==13.7.1
tikhub==1.13.0
websockets_proxy==0.1.2
python-dotenv
uvicorn